docs: correct inflation calculation in LIP-100 to account for compounding
The previous inflation formula assumed a flat annual rate and did not
account for compounding, resulting in a slight error when calculating
the maximum annual inflation under the 750,000 token ceiling. This
update revises the documentation to reflect the correct compounding
method. It also fixes some small typos in the LIP.

Co-authored-by: "Andrew W. Macpherson" <mail@awmacpherson.com>